cpu: disable gemm based implementations for threadpool
Gemm-based infrastructure requires a lot of changes to support
asynchronous runtime. However, it's impossible to disable asynchronous
runtime solely as threadpool is not avaialble at creation time through
conventional API. There's an option to introduce API to register a
threadpool before calling a primitive_desc creation function. If this
gets implemented, then introduced calls can be modified to check for
registered threadpool and if it is present, then check its flag to
decide if an implementation should be enabled/disabled. For the
simplicity this API is not a part of this commit.