cpu: x64: introduce get_wei_k_blk function
Also remove blocked layouts that use a K block size of 8, as they were introduced
for FP32 on AVX2 but do not appear to offer performance benefits over layouts
with a K block size of 16. We can always reintroduce them if they prove
beneficial in specific cases.

Also, enable the use of an N block size of 64 for FP32 on AVX2, as it provides
better performance.